[update] コンポーネントを共通化、ディレクトリを変更
