+++ src/App.tsx
@@ -3,6 +3,7 @@ import Footer from '@/Footer';
 import Header from '@/header/Header';
 import { Outlet } from 'react-router-dom';
 import theme from '@/theme/index';
+import VerificationProvider from '@/searchWindow/VerificationProvider';
 
 import { CssBaseline, ThemeProvider } from '@mui/material';
 
@@ -12,6 +13,7 @@ function App() {
         <ThemeProvider theme={theme}>
             <CssBaseline />
             <Header />
+            <VerificationProvider />
             <Outlet />
             <Footer />
         </ThemeProvider>
